pyproject.toml
[tool.poetry]
name = "movie-tv-show-tracker"
version = "0.1.0"
description = ""
authors = ["dmytro-dexter <dmytroboiko007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.27"
uvicorn = "^0.27.1"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
pydantic-settings = "^2.2.1"
black = "^24.3.0"
isort = "^5.13.2"
autoflake = "^2.3.1"
flake8 = "^7.0.0"
jinja2 = "^3.1.3"
uuid = "^1.30"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.7.0"}
python-multipart = "^0.0.9"
python-jose = "^3.3.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
